requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
toml==0.10.2
tqdm==4.66.1 